+++ src/event-log/event-log-level.ts
@@ -1,4 +1,4 @@
-import type { EventStream } from '../types/event-stream.js';
+import type { EventStream } from '../types/subscriptions.js';
 import type { ULIDFactory } from 'ulidx';
 import type { EventLog, GetEventsOptions } from '../types/event-log.js';
 import type { Filter, KeyValues } from '../types/query-types.js';
