+++ src/handlers/events-query.ts
@@ -1,7 +1,7 @@
 import type { DidResolver } from '../did/did-resolver.js';
 import type { EventLog } from '../types/event-log.js';
 import type { MethodHandler } from '../types/method-handler.js';
-import type { EventsQueryMessage, EventsQueryReply } from '../types/event-types.js';
+import type { EventsQueryMessage, EventsQueryReply } from '../types/events-types.js';
 
 import { EventsQuery } from '../interfaces/events-query.js';
 import { messageReplyFromError } from '../core/message-reply.js';
