+++ src/handlers/events-subscribe.ts
@@ -1,10 +1,10 @@
 import type { DidResolver } from '../did/did-resolver.js';
 import type EventEmitter from 'events';
-import type { EventStream } from '../types/event-stream.js';
+import type { EventStream } from '../types/subscriptions.js';
 import type { Filter } from '../types/query-types.js';
 import type { MessageStore } from '../types/message-store.js';
 import type { MethodHandler } from '../types/method-handler.js';
-import type { EventsSubscribeMessage, EventsSubscribeReply } from '../types/event-types.js';
+import type { EventsSubscribeMessage, EventsSubscribeReply } from '../types/events-types.js';
 
 import { EventsSubscribe } from '../interfaces/events-subscribe.js';
 import { Message } from '../core/message.js';
