+++ tests/handlers/events-subscribe.spec.ts
@@ -1,4 +1,4 @@
-import type { EventStream } from '../../src/types/event-stream.js';
+import type { EventStream } from '../../src/types/subscriptions.js';
 import type { DataStore, EventLog, GenericMessage, MessageStore } from '../../src/index.js';
 
 import { DidKeyResolver } from '../../src/did/did-key-resolver.js';
