+++ tests/interfaces/events-query.spec.ts
@@ -1,4 +1,4 @@
-import type { EventsQueryMessage } from '../../src/types/event-types.js';
+import type { EventsQueryMessage } from '../../src/types/events-types.js';
 import type { ProtocolsQueryFilter } from '../../src/types/protocols-types.js';
 import type { RecordsFilter } from '../../src/types/records-types.js';
