+++ tests/scenarios/end-to-end-tests.spec.ts
@@ -1,5 +1,5 @@
 import type { DerivedPrivateJwk } from '../../src/utils/hd-key.js';
-import type { EventStream } from '../../src/types/event-stream.js';
+import type { EventStream } from '../../src/types/subscriptions.js';
 import type { DataStore, EventLog, MessageStore, ProtocolDefinition, ProtocolsConfigureMessage, RecordsReadReply } from '../../src/index.js';
 
 import chaiAsPromised from 'chai-as-promised';
